requirements.txt
Flask==2.3.4
pyqrcode==1.2.1   # if using QR code generation
pypng==0.0.21     # for QR images
